Deduplicate key material accessors
