+++ crates/uselesskey-ecdsa/src/keypair.rs
@@ -7,10 +7,8 @@ use elliptic_curve::{
 };
 use rand_chacha10::ChaCha20Rng;
 use rand_core10::SeedableRng;
-use uselesskey_core::negative::CorruptPem;
-use uselesskey_core::sink::TempArtifact;
 use uselesskey_core::srp::keypair_material::Pkcs8SpkiKeyMaterial;
-use uselesskey_core::{Error, Factory};
+use uselesskey_core::Factory;
 
 use crate::EcdsaSpec;
 
@@ -147,170 +145,7 @@ impl EcdsaKeyPair {
         &self.label
     }
 
-    /// PKCS#8 DER-encoded private key bytes.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let der = kp.private_key_pkcs8_der();
-    /// assert!(!der.is_empty());
-    /// ```
-    pub fn private_key_pkcs8_der(&self) -> &[u8] {
-        self.inner.material.private_key_pkcs8_der()
-    }
-
-    /// PKCS#8 PEM-encoded private key.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let pem = kp.private_key_pkcs8_pem();
-    /// assert!(pem.starts_with("-----BEGIN PRIVATE KEY-----"));
-    /// ```
-    pub fn private_key_pkcs8_pem(&self) -> &str {
-        self.inner.material.private_key_pkcs8_pem()
-    }
-
-    /// SPKI DER-encoded public key bytes.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let der = kp.public_key_spki_der();
-    /// assert!(!der.is_empty());
-    /// ```
-    pub fn public_key_spki_der(&self) -> &[u8] {
-        self.inner.material.public_key_spki_der()
-    }
-
-    /// SPKI PEM-encoded public key.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let pem = kp.public_key_spki_pem();
-    /// assert!(pem.starts_with("-----BEGIN PUBLIC KEY-----"));
-    /// ```
-    pub fn public_key_spki_pem(&self) -> &str {
-        self.inner.material.public_key_spki_pem()
-    }
-
-    /// Write the PKCS#8 PEM private key to a tempfile and return the handle.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::Factory;
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::random();
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let temp = kp.write_private_key_pkcs8_pem().unwrap();
-    /// assert!(temp.path().exists());
-    /// ```
-    pub fn write_private_key_pkcs8_pem(&self) -> Result<TempArtifact, Error> {
-        self.inner.material.write_private_key_pkcs8_pem()
-    }
-
-    /// Write the SPKI PEM public key to a tempfile and return the handle.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::Factory;
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::random();
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let temp = kp.write_public_key_spki_pem().unwrap();
-    /// assert!(temp.path().exists());
-    /// ```
-    pub fn write_public_key_spki_pem(&self) -> Result<TempArtifact, Error> {
-        self.inner.material.write_public_key_spki_pem()
-    }
-
-    /// Produce a corrupted variant of the PKCS#8 PEM.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_core::negative::CorruptPem;
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let bad = kp.private_key_pkcs8_pem_corrupt(CorruptPem::BadHeader);
-    /// assert!(bad.contains("CORRUPTED"));
-    /// ```
-    pub fn private_key_pkcs8_pem_corrupt(&self, how: CorruptPem) -> String {
-        self.inner.material.private_key_pkcs8_pem_corrupt(how)
-    }
-
-    /// Produce a deterministic corrupted PKCS#8 PEM using a variant string.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::Factory;
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::random();
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let bad = kp.private_key_pkcs8_pem_corrupt_deterministic("corrupt:v1");
-    /// assert!(!bad.is_empty());
-    /// ```
-    pub fn private_key_pkcs8_pem_corrupt_deterministic(&self, variant: &str) -> String {
-        self.inner
-            .material
-            .private_key_pkcs8_pem_corrupt_deterministic(variant)
-    }
-
-    /// Produce a truncated variant of the PKCS#8 DER.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let truncated = kp.private_key_pkcs8_der_truncated(10);
-    /// assert_eq!(truncated.len(), 10);
-    /// ```
-    pub fn private_key_pkcs8_der_truncated(&self, len: usize) -> Vec<u8> {
-        self.inner.material.private_key_pkcs8_der_truncated(len)
-    }
-
-    /// Produce a deterministic corrupted PKCS#8 DER using a variant string.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use uselesskey_core::Factory;
-    /// # use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-    /// let fx = Factory::random();
-    /// let kp = fx.ecdsa("svc", EcdsaSpec::es256());
-    /// let bad = kp.private_key_pkcs8_der_corrupt_deterministic("corrupt:v1");
-    /// assert!(!bad.is_empty());
-    /// ```
-    pub fn private_key_pkcs8_der_corrupt_deterministic(&self, variant: &str) -> Vec<u8> {
-        self.inner
-            .material
-            .private_key_pkcs8_der_corrupt_deterministic(variant)
-    }
+    uselesskey_core::impl_pkcs8_spki_key_material_accessors!(inner.material);
 
     /// Return a valid (parseable) public key that does *not* match this private key.
     ///
